+++ kgql-core/src/commonMain/kotlin/com/codingfeline/kgql/core/KgqlRequestBody.kt
@@ -3,8 +3,8 @@ package com.codingfeline.kgql.core
 /**
  * Root interface for GraphQL request body
  */
-interface KgqlRequestBody<T> {
-    val operationName: String?
-    val query: String
-    val variables: T?
+public interface KgqlRequestBody<T> {
+    public val operationName: String?
+    public val query: String
+    public val variables: T?
 }