+++ kgql-core/src/commonMain/kotlin/com/codingfeline/kgql/core/KgqlResponse.kt
@@ -6,16 +6,16 @@ import kotlinx.serialization.Serializable
 /**
  * Root interface for GraphQL response
  */
-interface KgqlResponse<T> {
-    val data: T?
-    val errors: List<KgqlError>?
+public interface KgqlResponse<T> {
+    public val data: T?
+    public val errors: List<KgqlError>?
 }
 
 /**
  * GraphQL error
  */
 @Serializable
-data class KgqlError(
+public data class KgqlError(
     @SerialName("message") val message: String,
     @SerialName("locations") val locations: List<KgqlErrorLocation> = emptyList(),
     @SerialName("description") val description: String,
@@ -24,7 +24,7 @@ data class KgqlError(
 )
 
 @Serializable
-data class KgqlErrorLocation(
+public data class KgqlErrorLocation(
     @SerialName("line") val line: Int,
     @SerialName("column") val column: Int
 )